refatoração GlobalMapOverview.tsx
